SSCS-7229: MYA - Migrate unsubscribe hosted functionality from TYA to MYA
